refactor(task-engine): extract setup strategies behind runSetup dispatcher [SHR-191]
Move setupNew/Existing/None/Scratch into server/task-engine/setup/*.ts
behind a function-map dispatcher (Record<RunMode, SetupFn>); startTask's
mode switch collapses to await runSetup(task). SetupResult shape and all
git/fs argv preserved verbatim; no duplication found to extract.

Modular 04 step 04.2.
